LayoutTests/fast/canvas/gradient-into-p3-canvas-expected.txt

PASS false is false
PASS successfullyParsed is true

TEST COMPLETE
